refactor: move searched movies to its own component, delete unused import
